# openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MarketShield AI - Financial Market Surveillance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /detection/anomalies:
    get:
      operationId: listAnomalies
      tags: [detection]
      summary: ML anomaly detection results
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, under_review, confirmed, cleared]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Anomaly detection results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Anomaly"
...
  /investigations:
    get:
      operationId: listInvestigations
      tags: [investigations]
      summary: List all investigations
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, closed, escalated]
      responses:
        "200":
          description: Investigations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Investigation"
    post:
      operationId: createInvestigation
      tags: [investigations]
      summary: Create a new investigation case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestigationInput"
      responses:
        "201":
          description: Created investigation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investigation"
  /investigations/{id}:
    get:
      operationId: getInvestigation
      tags: [investigations]
      summary: Get investigation details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Investigation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investigation"
        "404":
          description: Not found
    patch:
      operationId: updateInvestigation
      tags: [investigations]
      summary: Update investigation status or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestigationUpdate"
      responses:
        "200":
          description: Updated investigation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investigation"
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List system alerts
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: resolved
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/resolve:
    post:
      operationId: resolveAlert
      tags: [alerts]
      summary: Mark alert as resolved
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List generated reports
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a new report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [reports]
      summary: Audit log entries
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
components:
  schemas:
...
    Anomaly:
      type: object
      required: [id, traderId, traderName, anomalyType, riskScore, detectedAt, status, modelUsed, confidence]
      properties:
        id:
          type: integer
        traderId:
          type: integer
        traderName:
          type: string
        anomalyType:
          type: string
        riskScore:
          type: number
        detectedAt:
          type: string
        status:
          type: string
          enum: [pending, under_review, confirmed, cleared]
        modelUsed:
          type: string
        confidence:
          type: number
        description:
          type: ["string", "null"]
...
    Investigation:
      type: object
      required: [id, caseNumber, title, status, priority, assignedTo, createdAt, updatedAt, traderName, riskScore, aiSummary]
      properties:
        id:
          type: integer
        caseNumber:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [open, in_progress, closed, escalated]
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignedTo:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        traderName:
          type: string
        riskScore:
          type: number
        aiSummary:
          type: string
        analystNotes:
          type: ["string", "null"]
        evidence:
          type: ["array", "null"]
          items:
            type: string
    InvestigationInput:
      type: object
      required: [title, traderName, riskScore, priority]
      properties:
        title:
          type: string
        traderName:
          type: string
        riskScore:
          type: number
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignedTo:
          type: string
    InvestigationUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, closed, escalated]
        analystNotes:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignedTo:
          type: string
    Alert:
      type: object
      required: [id, alertType, title, message, severity, isResolved, createdAt, relatedEntity]
      properties:
        id:
          type: integer
        alertType:
          type: string
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        isResolved:
          type: boolean
        createdAt:
          type: string
        relatedEntity:
          type: string
        resolvedAt:
          type: ["string", "null"]
...
    Report:
      type: object
      required: [id, title, reportType, generatedAt, generatedBy, status, period]
      properties:
        id:
          type: integer
        title:
          type: string
        reportType:
          type: string
          enum: [regulatory, trend_analysis, investigation_summary, compliance]
        generatedAt:
          type: string
        generatedBy:
          type: string
        status:
          type: string
          enum: [generating, ready, failed]
        period:
          type: string
        downloadUrl:
          type: ["string", "null"]
    ReportInput:
      type: object
      required: [title, reportType, period]
      properties:
        title:
          type: string
        reportType:
          type: string
          enum: [regulatory, trend_analysis, investigation_summary, compliance]
        period:
          type: string
    AuditLog:
      type: object
      required: [id, action, userId, userName, timestamp, resource, details]
      properties:
        id:
          type: integer
        action:
          type: string
        userId:
          type: integer
        userName:
          type: string
        timestamp:
          type: string
        resource:
          type: string
        details:
          type: string